Inscrições das atividades

+++ database/migrations/2019_12_06_155900_create_ext_atv_turmas_table.php
@@ -24,7 +24,7 @@ class CreateExtAtvTurmasTable extends Migration
             $table->dateTime('dia_libera');
             $table->dateTime('dia_bloqueia');
             $table->string('user');
-            $table->unsignedInteger('ext_atvs_id');            
+            $table->unsignedBigInteger('ext_atvs_id');            
             $table->foreign('ext_atvs_id')
                 ->references('id')
                 ->on('ext_atvs')
